refactor: tetra.core.client module
