+++ tetra/__init__.py
@@ -1,12 +1,11 @@
 # Import all the packahges here
 
-from .client import RemoteExecutionClient, remote
 from . import remote_execution_pb2, remote_execution_pb2_grpc
 from .core.resources.resource_manager import ResourceManager
 from .core.resources.serverless import ServerlessResource
+from .core.client.client_manager import remote
 
 __all__ = [
-    "RemoteExecutionClient",
     "remote",
     "remote_execution_pb2",
     "remote_execution_pb2_grpc",
